Add files via upload
Day 4 and Day 5 Task

One of Ingressive for Good’s co-founders is sponsoring a design contest for a family business. All God’s Children Learning Center is daycare center based in Gautier, Mississippi USA and the founder will review the branding package (logo, business cards and other brand identity assets) submissions to pick the winner. The prize for the finalist is N50,000.00